Convert sensor data from ints to floats
Originally all fields were ints, but now with the addition of
distance and speed there are also floats. Rather than use
separate float / int sensor values it was assumed simpler to
convert them all to floats and update the tests.

This also adds a test for the DistanceMeters field in the the
TCX file, but not for speed which will probably be removed
later.

+++ Cyclismo/Cyclismo/src/main/java/org/cowboycoders/cyclismo/io/file/CsvTrackWriter.java
@@ -186,25 +186,25 @@ public class CsvTrackWriter implements TrackFormatWriter {
         if (sensorDataSet.hasDistance()) {
           SensorData sensorData = sensorDataSet.getDistance();
           if (sensorData.hasValue() && sensorData.getState() == Sensor.SensorState.SENDING) {
-            distance = Double.toString(sensorData.getValue());
+            distance = Float.toString(sensorData.getValue());
           }
         }
         if (sensorDataSet.hasPower()) {
           SensorData sensorData = sensorDataSet.getPower();
           if (sensorData.hasValue() && sensorData.getState() == Sensor.SensorState.SENDING) {
-            power = Double.toString(sensorData.getValue());
+            power = Float.toString(sensorData.getValue());
           }
         }
         if (sensorDataSet.hasCadence()) {
           SensorData sensorData = sensorDataSet.getCadence();
           if (sensorData.hasValue() && sensorData.getState() == Sensor.SensorState.SENDING) {
-            cadence = Double.toString(sensorData.getValue());
+            cadence = Float.toString(sensorData.getValue());
           }
         }
         if (sensorDataSet.hasHeartRate()) {
           SensorData sensorData = sensorDataSet.getHeartRate();
           if (sensorData.hasValue() && sensorData.getState() == Sensor.SensorState.SENDING) {
-            heartRate = Double.toString(sensorData.getValue());
+            heartRate = Float.toString(sensorData.getValue());
           }
         }
       }
